백준 2420 : 오류 수정

+++ src/algorithm/SafariWorld2420.java
@@ -6,10 +6,11 @@ public class SafariWorld2420 {
     public static void main(String[] args){
         Scanner sc = new Scanner(System.in);
 
-        int A = sc.nextInt();
-        int B = sc.nextInt();
-        int Sub = Math.abs(A-B);
+        long A = sc.nextLong();
+        long B = sc.nextLong();
+        long Sub = Math.abs(A-B);
 
         System.out.println(Sub);
+
     }
 }
